+++ examples/xApp/go/FlexPolicy/utils/sm/slice/slice.go
@@ -36,10 +36,10 @@ func (c SLICECallback) Handle(ind xapp.Swig_slice_ind_msg_t) {
 
 // E2NodeSliceStatsMap is a map of E2Node ID to SliceStatsDict
 type E2NodeSliceStatsMap struct {
-	Mcc              int16          `json:"mcc,omitempty"`
-	Mnc              int16          `json:"mnc,omitempty"`
-	NbId             int16          `json:"nb_id,omitempty"`
-	CuDuId           int16          `json:"cu_du_id,omitempty"`
+	Mcc              uint16         `json:"mcc,omitempty"`
+	Mnc              uint16         `json:"mnc,omitempty"`
+	NbId             uint32         `json:"nb_id,omitempty"`
+	CuDuId           uint32         `json:"cu_du_id,omitempty"`
 	RanType          string         `json:"ran_type,omitempty"`
 	E2NodeSliceStats SliceStatsDict `json:"slice_stats,omitempty"`
 }
@@ -101,10 +101,10 @@ var SliceStats []E2NodeSliceStatsMap
 var Mutex sync.Mutex
 
 type E2NodeId struct {
-	Mcc     int16
-	Mnc     int16
-	NbId    int16
-	CuDuId  int16
+	Mcc     uint16
+	Mnc     uint16
+	NbId    uint32
+	CuDuId  uint32
 	RanType string
 }
 
@@ -385,11 +385,11 @@ func SliceIndToDictJSON(ind xapp.Swig_slice_ind_msg_t) {
 	Mutex.Lock()
 
 	// get E2node details
-	NbId := int16(ind.GetId().GetNb_id().GetNb_id())
+	NbId := uint32(ind.GetId().GetNb_id().GetNb_id())
 	// CuDuId := ind.GetId().GetCu_du_id() // TODO: not yet supported, maybe needs swig extension
 	RanType := xapp.Get_e2ap_ngran_name(ind.GetId().GetXtype())
-	Mcc := int16(ind.GetId().GetPlmn().GetMcc())
-	Mnc := int16(ind.GetId().GetPlmn().GetMnc())
+	Mcc := ind.GetId().GetPlmn().GetMcc()
+	Mnc := ind.GetId().GetPlmn().GetMnc()
 
 	// Find the entry in the global structure SliceStats if it exists or create a new entry
 	entryFound := false
